refactor: 拆分guide界面hooks到不同模块
